+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Next
 - Added telescope (gated to admin only) to better debug.
+- Moved validation of new places into its own `FormRequest`. We now validate URL, email, phone number, required fields and conditional required fields based on other fields.
 
 # 1.5.0 (2020-03-22)
 - Divided js and css resources in two parts (frontend and backend), and extracted vendor files.